+++ capp/Cargo.toml
@@ -1,6 +1,73 @@
 [package]
 name = "capp"
-version = "0.1.0"
+version = "0.4.4"
 edition = "2021"
+license = "MIT"
+authors = ["oiwn"]
+description = "Common things i use to build Rust CLI tools for web crawlers."
+homepage = "https://github.com/oiwn/capp-rs"
+repository = "https://github.com/oiwn/capp-rs"
+readme = "README.md"
+keywords = ["web-crawler", "async", "executor"]
+categories = ["asynchronous", "web-programming", "concurrency"]
+exclude = [
+    "tmp/*",
+    # "benches/*",
+    "html/*",
+    ".github/*",
+    ".gitignore",
+    ".tmuxp.yaml",
+    "notes.org"
+]
+
+[lib]
+name = "capp"
+path = "src/lib.rs"
+doctest = false
 
 [dependencies]
+serde.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+anyhow.workspace = true
+tracing.workspace = true
+tracing-subscriber.workspace = true
+# derive_builder.workspace = true
+
+async-trait = { version = "0.1" }
+backoff = { version = "0.4", optional = true, features = ["tokio"] }
+derive_builder = { version = "0.20" }
+reqwest = { version = "0.12", features = ["gzip", "rustls-tls", "json"], optional = true }
+# serde = { version = "1.0", features = ["derive"] }
+# serde_json = { version = "1.0" }
+serde_yaml = "0.9"
+# thiserror = { version = "1" }
+tokio = { version = "1.41", features = ["full"] }
+uuid = { version = "1.11", features = ["v4", "serde"] }
+rustis = { version = "0.13", features = ["tokio-runtime"], optional = true }
+# tracing = "0.1"
+# tracing-subscriber = "0.3"
+# anyhow = "1"
+tracing-futures = "0.2"
+indexmap = "2.6"
+url = "2.5"
+regex = "1.11"
+rand = { version = "0.8", optional = true } 
+
+[dev-dependencies]
+capp = { path = ".", features = ["http", "healthcheck", "redis"] }
+hyper = { version = "1.5", features = ["server", "http1"] }
+http-body-util = "0.1"
+bytes = "1.6"
+pin-project-lite = "0.2"
+dotenvy = "0.15"
+scraper = "0.21"
+rand = "0.8"
+md5 = "0.7"
+url = "2.5"
+base64 = "0.22"
+
+[features]
+http = ["dep:backoff", "dep:reqwest", "dep:rand"]
+healthcheck = ["dep:reqwest"]
+redis = ["dep:rustis"]
