+++ capp/src/manager.rs
