+++ capp/src/manager/computation.rs
