+++ capp/src/manager/stats.rs
