+++ capp/src/manager/workers_manager.rs
